moved paths struct
