+++ bundle/config/paths/paths.go
@@ -1,4 +1,4 @@
-package resources
+package paths
 
 import (
 	"fmt"
